+++ MinimOsd_Extra/0-build.sh
@@ -16,6 +16,6 @@ make_one(){
 
 
 #make_one 'UAVTALK'
-make_one 'MAVLINK'
-#make_one 'MWII'
+#make_one 'MAVLINK'
+make_one 'MWII'
 #make_one 'LTM'